requirements.txt
fastapi
uvicorn
moviepy
python-dotenv
requests
edge-tts
google-genai
pillow